backend/pyproject.toml
[project]
name = "yolo-deck-backend"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "docker>=7.0.0",
    "websockets>=13.0",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "gitpython>=3.1.40",
]
